requirements.txt
streamlit==1.28.1
pandas==2.1.3
kiteconnect==4.2.0
python-dotenv==1.0.0 